requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
aiosqlite>=0.20.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
httpx>=0.28.0
python-multipart>=0.0.12
bcrypt>=4.2.0
python-jose[cryptography]>=3.3.0
email-validator>=2.2.0
jinja2>=3.1.4
aiofiles>=24.1.0
python-dotenv>=1.0.1
itsdangerous>=2.2.0
